Add weappon pickup, and mecanism for pickups in general.
ActiveInventory component is responisble for handling the
active invnentory of the player, for now it has only one
slot, for the weapon.
AEquippable has 3 callbacks that are called when the weapon
gets equipped in an ActiveInventoryComponent.
To pick up a weapon, for now a player walks over it. This
is FOR NOW not done via the Interactable interface, but
this should be change ASAP.
Also a weapon should be attached to the player upon pickup.