add 3rd person camera
Added VERY basic 3rd person camera, visual is is not an improvement
over the old one, but the camera is now it it's own class
ThirdPersonCameraActor, that is spawned by the controller, and
is attached to the pawn on Possess.
It also seems that at lest for now it's not necesearry to overload
PlayerCameraManager, as the functionaly needed is present (I just
need to figure it out).