Add working prototype for weapon equipping
Now when AWeapon is touched (and in the future Interact()ed with)
it is equipped (ie, attached to the player pawn to a socket,
added to the ActiveInventoryComponent, and notified that it is
being equipped).
Note that to get acess to GetMesh() the Equippable class has
to assume that the actor equipping is a child of ACharacter.